pyproject.toml
[project]
name = "projeto-calculo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=3.1.1",
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sympy>=1.14.0",
    "uvicorn>=0.38.0",
]
